backend/requirements/base.txt
# IntelliPDF Backend - Base Dependencies
# Python 3.11+

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Async Support
anyio==4.2.0
httpx==0.26.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis[hiredis]==5.0.1
aioredis==2.0.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# AI/ML
langchain==0.1.0
langchain-openai==0.0.2
openai==1.7.2
tiktoken==0.5.2

# Vector Database
chromadb==0.4.22
sentence-transformers==2.3.1

# PDF Processing
pypdf2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.8
pdf2image==1.16.3
pillow==10.2.0

# Text Processing
spacy==3.7.2
nltk==3.8.1
python-docx==1.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3

# Date/Time
python-dateutil==2.8.2
